requirements.txt
streamlit>=1.28.0
langextract>=0.1.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
pillow>=9.0.0
python-dotenv>=0.19.0